Backend/requirements.txt
numpy==1.24.4
pillow==10.4.0
